pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finagle"
version = "0.1.0"
description = "API-first personal finance app for Australian CGT calculations"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "python-multipart>=0.0.18",
    "openpyxl>=3.1.5",
    "python-calamine>=0.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.28",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.setuptools.packages.find]
include = ["app*"]
